backend/requirements.txt
fastapi
uvicorn
yfinance
ta
langchain
langchain-openai
python-multipart
psycopg2-binary
sqlalchemy
alembic
